`feat: Implement 'fuck' command to curse other players`
This commit introduces the 'fuck' command, which allows players with a dice roll of 1 or 2 to curse the next player who rolls a number equal to or lower than their roll.  This adds a new layer of interaction and strategy to the game.

The commit also includes tests for the new command and modifications to the dice handler to trigger the curse when appropriate.